+++ src/utils.rs
@@ -1,9 +1,8 @@
 use crate::jwt::{decode_jwt, Claims};
 use crate::response::error::MomentoError;
 
-pub fn is_ttl_valid(ttl: &u32) -> Result<(), MomentoError> {
-    // max_ttl will be 4294967 since 2^32 / 1000 = 4294967.296
-    let max_ttl = u32::MAX / 1000 as u32;
+pub fn is_ttl_valid(ttl: &u64) -> Result<(), MomentoError> {
+    let max_ttl = u64::MAX / 1000 as u64;
     if *ttl > max_ttl {
         return Err(MomentoError::InvalidArgument(format!(
             "TTL provided, {}, needs to be less than the maximum TTL {}",
